+++ bciers/apps/registration1/app/components/userOperators/UserOperatorForm.tsx
@@ -11,7 +11,7 @@ import {
 import { userOperatorUiSchema } from "@/app/utils/jsonSchema/userOperator";
 import { actionHandler } from "@/app/utils/actions";
 import { useSession } from "next-auth/react";
-import MultiStepFormBase from "@/app/components/form/MultiStepFormBase";
+import MultiStepFormBase from "@bciers/components/form/MultiStepFormBase";
 import { UserOperatorFormData } from "@/app/components/form/formDataTypes";
 import Note from "@/app/components/datagrid/Note";
 import { Status } from "@/app/utils/enums";
