+++ bciers/libs/shared/components/src/lib/form/components/MultiStepButtons.test.tsx
@@ -1,4 +1,4 @@
-import MultiStepButtons from "@/app/components/form/MultiStepButtons";
+import MultiStepButtons from "@bciers/components/form/components/MultiStepButtons";
 import { render, screen } from "@testing-library/react";
 import { describe, expect } from "vitest";
 import React from "react";
