+++ bciers/libs/shared/components/src/lib/form/index.ts
@@ -1,4 +1,3 @@
 export { default as FormBase } from "./FormBase";
-export { default as SingleStepTaskListForm } from "./SingleStepTaskListForm";
-export { default as SingleStepTaskListButtons } from "./SingleStepTaskListButtons";
-export { default as TaskList } from "./TaskList";
+export { default as MultiStepAccordion } from "./MultiStepAccordion";
+export { default as MultiStepFormBase } from "./MultiStepFormBase";
